physics_environment/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
numpy==1.26.4
scipy==1.12.0
protobuf==4.25.3
